:cherry_blossom: feat: add delete success message #11
Contextつくって、AppにMessageを配置
