+++ src/context/SuccessMessageContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const SuccessMessageContext = createContext(null);
